requirements.txt
apiview==1.3.25
appdirs==1.4.4
asgiref==3.3.1
bar==0.2.1
certifi==2020.12.5
chainmap==1.0.3
chardet==4.0.0
combomethod==1.0.12
csrf==0.1b1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-core==1.4.1
django-heroku==0.3.1
django-import-export==2.5.0
django-toolbelt==0.0.1
djangorestframework==3.12.4
echarts==0.0.0
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
Jinja2==2.11.3
location==0.0.7
MarkupPy==1.14
MarkupSafe==1.1.1
null==0.6.1
nulltype==2.3.1
odfpy==1.4.1
openpyxl==3.0.7
options==1.4.10
peppercorn==0.6
pipenv==2020.11.15
prettytable==2.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyecharts==1.9.0
PyMySQL==1.0.2
python-dateutil==1.5
pythonw==3.0.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
six==1.12.0
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
